third_party/WebKit/LayoutTests/external/wpt/webrtc/RTCPeerConnection-setRemoteDescription-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Error: assert_equals: expected "stable" but got "closed"
PASS setRemoteDescription with invalid type and invalid SDP should reject with TypeError
FAIL setRemoteDescription() with invalid SDP and stable state should reject with InvalidStateError assert_throws: function "() => { throw e }" threw object "OperationError: Failed to parse SessionDescription. invalid Expect line: v=" that is not a DOMException InvalidStateError: property "code" is equal to 0, expected 11
FAIL Negotiation should fire signalingsstate events assert_equals: expected "stable" but got "have-local-offer"
FAIL Calling setRemoteDescription() again after one round of remote-offer/local-answer should succeed promise_test: Unhandled rejection with value: object "InvalidAccessError: Failed to set remote offer sdp: The order of m-lines in subsequent offer doesn't match order from previous offer/answer."
FAIL Switching role from offerer to answerer after going back to stable state should succeed promise_test: Unhandled rejection with value: object "TypeError: Cannot read property 'sdp' of undefined"
Harness: the test ran to completion.
